Marae & Lethice level adjustments for New Game Plus
And some rearrangement of Marae code so the conditional alteration after initial declaration gets moved down.

In addition, Marae's level goes up by 60 as opposed to 30 per tier of New Game+, going up to a max of 339!

Her level defaults to 85 but is lowered to 75 if you overloaded the factories, and 99 if you planted the acorn.